Update to main.js
The game function is still not properly operable but I changed a few of the variables to be constant as the don't change. I also updated the layout of my js file by putting on the variables up top and functions below for easier readability and to ensure my code functions properly. Last change is that I consolidated my playRound() function to take up less space and get rid of repeating lines of code.