Update to main.js - Score added
I successfully added the score and amountOfRounds features into the game. I previously had a separate function to run multiple games and return an outcome. Instead, I figured out that if you just place another if-then in the original ones after the score increase is added, it can check to see if the amount of rounds requested is met.